Refactor `_init_trace` out of `mcmc` module
This uncouples several things:
* `_init_trace` is now independent of abstract step-methods
* `mcmc` is now unaware of `NDArray`
* code for population-sampling can now be extracted from `mcmc`
