+++ pymc/tests/backends/test_base.py
@@ -14,6 +14,9 @@
 import numpy as np
 import pytest
 
+import pymc as pm
+
+from pymc.backends import _init_trace
 from pymc.backends.base import _choose_chains
 
 
@@ -31,3 +34,22 @@ def test_choose_chains(n_points, tune, expected_length, expected_n_traces):
     traces, length = _choose_chains([trace_0, trace_1, trace_2], tune=tune)
     assert length == expected_length
     assert expected_n_traces == len(traces)
+
+
+class TestInitTrace:
+    def test_init_trace_continuation_unsupported(self):
+        with pm.Model() as pmodel:
+            A = pm.Normal("A")
+            B = pm.Uniform("B")
+            strace = pm.backends.ndarray.NDArray(vars=[A, B])
+            strace.setup(10, 0)
+            strace.record({"A": 2, "B_interval__": 0.1})
+            assert len(strace) == 1
+            with pytest.raises(ValueError, match="Continuation of traces"):
+                _init_trace(
+                    expected_length=20,
+                    stats_dtypes=pm.Metropolis().stats_dtypes,
+                    chain_number=0,
+                    trace=strace,
+                    model=pmodel,
+                )
